requirements.txt
networkx~=2.6.3
numpy~=1.21.4
scipy~=1.7.3
torch~=1.10.0
pyro-ppl

setuptools==59.4.0
wheel~=0.37.0

docutils~=0.17.1
py~=1.11.0
Pillow~=8.4.0
pip~=20.0.2
attrs~=21.2.0
tqdm~=4.62.3
pytz~=2021.3
toml~=0.10.2
requests~=2.26.0
Babel~=2.9.1
pytest~=6.2.5
matplotlib~=3.5.0
tomli~=1.2.2
six~=1.16.0
urllib3~=1.26.7
MarkupSafe~=2.0.1
Sphinx~=4.3.1
packaging~=21.3
Jinja2~=3.0.3
imagesize~=1.3.0
snowballstemmer~=2.2.0
Pygments~=2.10.0
pluggy~=1.0.0
iniconfig~=1.1.1
anytree~=2.8.0
certifi~=2021.10.8
python-dateutil~=2.8.2
idna~=3.3
alabaster~=0.7.12
fonttools
pyparsing~=3.0.6
cycler~=0.11.0
kiwisolver

cffi~=1.15.0
tzlocal~=4.1
pycparser~=2.21

fire~=0.4.0
mpmath~=1.2.1